+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 mobile_menu Changelog
 =====================
 
+Version 2.1.0 stable (2015-09-23)
+---------------------------------
+
+### Improved
+- The trigger has now the active class when menu is open
+- Added the option to use custom close buttons inside the menu
+
+
 Version 2.0.0 stable (2015-09-19)
 ---------------------------------